Implement group code sharing ============================
Each installation has a dedicated group code. This code never changes.
It is stored in the local storage of the browser. The key is group.
The user is shown a screen with a choice:
- Create new group
- Join existing group

If the user chooses to create a new group, two things happen:
- The user's decision is stored in the local storage of the browser. with
  the key as 'groupSelected' and the value as 'group'
- The group code will be the code generated when the user first
  installed the extension.

If the user chooses to join an existing group, the user is asked to enter
an eight character code they would have received from another user.
- The user's decision is stored in the local storage of the browser. with
  the key as 'groupSelected' and the value as 'groupOther'
- The code the user entered is stored in the local storage of the browser.
  with the key as 'groupOther' and the value as the code the user entered.

This perhaps over engineered flag arrangement allows for the following:
- It allows the Flats component to work with both user owned and joined
groups.
- It always allows a user to get back to their own list even if they
joined another group.